+++ src/index.ts
@@ -632,8 +632,8 @@ export class Token extends AstElement {
 
 	/** @private */
 	inTableAttrs(): 1 | 2 | false {
-		return this.closest('table-attrs,ext')?.type === 'table-attrs' && (
-			this.closest('table-attrs,arg,magic-word,template')?.is<AttributesToken>('table-attrs')
+		return this.isInside('table-attrs') && (
+			this.closest('table-attrs,arg,parameter')?.is<AttributesToken>('table-attrs')
 				? 2
 				: 1
 		);
@@ -641,9 +641,7 @@ export class Token extends AstElement {
 
 	/** @private */
 	inHtmlAttrs(): 1 | 2 | false {
-		return this.closest('html-attrs,ext')?.is<AttributesToken>('html-attrs')
-			? 2
-			: this.inTableAttrs();
+		return this.isInside('html-attrs') ? 2 : this.inTableAttrs();
 	}
 
 	/** @private */